KoGPT2_chatbot_requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aniso8601==9.0.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.6.0
bidict==0.21.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cortex-python==6.0.3
cryptography==3.4.7
cuid==0.3
dataclasses==0.8
Deprecated==1.2.12
dill==0.3.4
enum-compat==0.0.3
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
flask-ngrok==0.0.25
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.0
fsspec==2021.6.0
future==0.18.2
gevent==21.1.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
greenlet==1.1.0
grpcio==1.38.0
huggingface-hub==0.0.8
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
jwcrypto==1.0
konlpy==0.5.2
Korpora==0.2.0
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
object-detection==0.1
packaging==20.9
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.2
psutil==5.8.0
py-hanspell==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-engineio==4.2.0
python-jwt==3.3.0
python-socketio==5.3.0
pytorch-lightning==1.2.10
pytz==2021.1
PyYAML==5.3.1
ratsnlp==0.0.9998
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rsa==4.7.2
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tokenizers==0.10.3
torch==1.9.0
torch-model-archiver==0.4.2
torchmetrics==0.2.0
torchserve==0.2.0
torchvision==0.10.0
tqdm==4.61.0
transformers==4.5.1
tweepy==4.3.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
